[RLlib] Clean up mixins.
