+++ rllib/policy/tf_policy.py
@@ -1204,100 +1204,3 @@ class TFPolicy(Policy):
             feed_dict[self._seq_lens] = train_batch[SampleBatch.SEQ_LENS]
 
         return feed_dict
-
-
-@DeveloperAPI
-class LearningRateSchedule:
-    """Mixin for TFPolicy that adds a learning rate schedule."""
-
-    @DeveloperAPI
-    def __init__(self, lr, lr_schedule):
-        self._lr_schedule = None
-        if lr_schedule is None:
-            self.cur_lr = tf1.get_variable("lr", initializer=lr, trainable=False)
-        else:
-            self._lr_schedule = PiecewiseSchedule(
-                lr_schedule, outside_value=lr_schedule[-1][-1], framework=None
-            )
-            self.cur_lr = tf1.get_variable(
-                "lr", initializer=self._lr_schedule.value(0), trainable=False
-            )
-            if self.framework == "tf":
-                self._lr_placeholder = tf1.placeholder(dtype=tf.float32, name="lr")
-                self._lr_update = self.cur_lr.assign(
-                    self._lr_placeholder, read_value=False
-                )
-
-    @override(Policy)
-    def on_global_var_update(self, global_vars):
-        super(LearningRateSchedule, self).on_global_var_update(global_vars)
-        if self._lr_schedule is not None:
-            new_val = self._lr_schedule.value(global_vars["timestep"])
-            if self.framework == "tf":
-                self.get_session().run(
-                    self._lr_update, feed_dict={self._lr_placeholder: new_val}
-                )
-            else:
-                self.cur_lr.assign(new_val, read_value=False)
-                # This property (self._optimizer) is (still) accessible for
-                # both TFPolicy and any TFPolicy_eager.
-                self._optimizer.learning_rate.assign(self.cur_lr)
-
-    @override(TFPolicy)
-    def optimizer(self):
-        return tf1.train.AdamOptimizer(learning_rate=self.cur_lr)
-
-
-@DeveloperAPI
-class EntropyCoeffSchedule:
-    """Mixin for TFPolicy that adds entropy coeff decay."""
-
-    @DeveloperAPI
-    def __init__(self, entropy_coeff, entropy_coeff_schedule):
-        self._entropy_coeff_schedule = None
-        if entropy_coeff_schedule is None:
-            self.entropy_coeff = get_variable(
-                entropy_coeff, framework="tf", tf_name="entropy_coeff", trainable=False
-            )
-        else:
-            # Allows for custom schedule similar to lr_schedule format
-            if isinstance(entropy_coeff_schedule, list):
-                self._entropy_coeff_schedule = PiecewiseSchedule(
-                    entropy_coeff_schedule,
-                    outside_value=entropy_coeff_schedule[-1][-1],
-                    framework=None,
-                )
-            else:
-                # Implements previous version but enforces outside_value
-                self._entropy_coeff_schedule = PiecewiseSchedule(
-                    [[0, entropy_coeff], [entropy_coeff_schedule, 0.0]],
-                    outside_value=0.0,
-                    framework=None,
-                )
-
-            self.entropy_coeff = get_variable(
-                self._entropy_coeff_schedule.value(0),
-                framework="tf",
-                tf_name="entropy_coeff",
-                trainable=False,
-            )
-            if self.framework == "tf":
-                self._entropy_coeff_placeholder = tf1.placeholder(
-                    dtype=tf.float32, name="entropy_coeff"
-                )
-                self._entropy_coeff_update = self.entropy_coeff.assign(
-                    self._entropy_coeff_placeholder, read_value=False
-                )
-
-    @override(Policy)
-    def on_global_var_update(self, global_vars):
-        super(EntropyCoeffSchedule, self).on_global_var_update(global_vars)
-        if self._entropy_coeff_schedule is not None:
-            new_val = self._entropy_coeff_schedule.value(global_vars["timestep"])
-            if self.framework == "tf":
-                self.get_session().run(
-                    self._entropy_coeff_update,
-                    feed_dict={self._entropy_coeff_placeholder: new_val},
-                )
-            else:
-                self.entropy_coeff.assign(new_val, read_value=False)
